Initial Commit
Game Maker was built as a part of course work for Object Oriented
Software Development. It allows naive users to create games such as
Pacman, Breakout, Frogger, etc.
A simple demo can be viewed at
https://www.youtube.com/watch?v=JSWU8H8nCLc
The highlights of the project were:
1. Developed in Java, used MVC, Observer and Command design patterns.
2. Improvised the feature of the application every week and utilized
iterative design of agile software development.
3.  Worked in a team of 5 and it keeps on switching every week to get
the core idea of working with different people.
4.  The code base to work on keeps on switching every week so we learn
to make our code more maintainable and extensible.